requirements.txt
google-generativeai
openpyxl
grpcio==1.60.1